refactor: split models into their own files
