+++ src/augmenters/ldap_augmenter.rs
@@ -9,7 +9,7 @@ use serde::{Deserialize, Serialize};
 use tracing::{debug, info, warn};
 
 use crate::augmenters::Augmenter;
-use crate::models::User;
+use crate::models::user::User;
 
 /// Configuration required to connect to LDAP and fetch user roles.
 #[derive(Deserialize, Serialize, JsonSchema, Debug, Hash, PartialEq, Eq, Clone)]
