+++ src/models/mod.rs
@@ -1,0 +1,2 @@
+pub mod token;
+pub mod user;
