+++ src/store/store.rs
@@ -5,7 +5,7 @@ use tracing::{debug, error, info};
 
 use super::{mongodb_store::MongoDBStore, no_store::NoStore};
 use crate::config::{StoreBackend, StoreConfig};
-use crate::models::{Token, User};
+use crate::models::{token::Token, user::User};
 use crate::providers::Provider;
 
 /// The Store trait abstracts token storage (add, retrieve, delete).
